fix in MM PrepRawData

+++ MuonSpectrometer/MuonReconstruction/MuonRecEvent/MuonPrepRawData/MuonPrepRawData/MMPrepData.h
@@ -81,10 +81,10 @@ namespace Muon
     int time() const;
 
     /** @brief Returns the ADC counts */
-    int charge() const;
+    double charge() const;
 
     /** @brief Returns the microTPC angle */
-    int angle() const;
+    double angle() const;
 
     /** @brief Returns the microTPC chisq Prob. */
     int chisqProb() const;
@@ -130,12 +130,12 @@ namespace Muon
     return m_time;
   }
   
-  inline int MMPrepData::charge() const 
+  inline double MMPrepData::charge() const 
   {
     return m_charge;
   }
 
-  inline int MMPrepData::angle() const 
+  inline double MMPrepData::angle() const 
   {
     return m_angle;
   }
